+++ include/cRIO/NiStatus.h
@@ -39,16 +39,6 @@ namespace cRIO {
  */
 const char* NiStatus(int32_t status);
 
-/**
- * Log NiFpga call error or warning.
- *
- * @param msg additional message (can be __PRETTY_FUNCTION__)
- * @param status NiFpga returned status. < 0 for errors, > 0 for warning
- *
- * @return status (so it can be reused)
- */
-int32_t NiReportErrorWarning(const char* msg, int32_t status);
-
 }  // namespace cRIO
 }  // namespace LSST