+++ doc/version-history.rst
@@ -6,6 +6,7 @@ v1.10.1
 -------
 
 * Set and read DCA Gain commands (ILC code 73 and 74)
+* Add NiWarning class, improve Ni error reporting
 
 v1.10.0
 -------